requirements.txt
Flask==3.0.2
Flask-RESTful==0.3.10
python-decouple==3.8